guesser interface
