--- source/Patches/CrewmateRoles/VigilanteMod/ShowHideButtonsVigi.cs
@@ -1,68 +1,0 @@
-﻿using HarmonyLib;
-using TownOfUs.Roles;
-using UnityEngine.UI;
-
-namespace TownOfUs.CrewmateRoles.VigilanteMod
-{
-    [HarmonyPatch(typeof(MeetingHud), nameof(MeetingHud.Confirm))]
-    public class ShowHideButtonsVigi
-    {
-        public static void HideButtonsVigi(Vigilante role)
-        {
-            foreach (var (_, (cycleBack, cycleForward, guess, guessText)) in role.Buttons)
-            {
-                if (cycleBack == null || cycleForward == null) continue;
-                cycleBack.SetActive(false);
-                cycleForward.SetActive(false);
-                guess.SetActive(false);
-                guessText.gameObject.SetActive(false);
-
-                cycleBack.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-                cycleForward.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-                guess.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-                role.GuessedThisMeeting = true;
-            }
-        }
-
-        public static void HideSingle(
-            Vigilante role,
-            byte targetId,
-            bool killedSelf
-        )
-        {
-            if (
-                killedSelf ||
-                role.RemainingKills == 0 ||
-                (!CustomGameOptions.VigilanteMultiKill)
-            ) HideButtonsVigi(role);
-            else HideTarget(role, targetId);
-        }
-        public static void HideTarget(
-            Vigilante role,
-            byte targetId
-        )
-        {
-
-            var (cycleBack, cycleForward, guess, guessText) = role.Buttons[targetId];
-            if (cycleBack == null || cycleForward == null) return;
-            cycleBack.SetActive(false);
-            cycleForward.SetActive(false);
-            guess.SetActive(false);
-            guessText.gameObject.SetActive(false);
-
-            cycleBack.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-            cycleForward.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-            guess.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-            role.Buttons[targetId] = (null, null, null, null);
-            role.Guesses.Remove(targetId);
-        }
-
-
-        public static void Prefix(MeetingHud __instance)
-        {
-            if (!PlayerControl.LocalPlayer.Is(RoleEnum.Vigilante)) return;
-            var retributionist = Role.GetRole<Vigilante>(PlayerControl.LocalPlayer);
-            if (!CustomGameOptions.VigilanteAfterVoting) HideButtonsVigi(retributionist);
-        }
-    }
-}
