--- source/Patches/Modifiers/AssassinMod/NotVote.cs
@@ -1,18 +1,0 @@
-using HarmonyLib;
-using TownOfUs.Roles.Modifiers;
-
-namespace TownOfUs.Modifiers.AssassinMod
-{
-    [HarmonyPatch(typeof(MeetingHud), nameof(MeetingHud.VotingComplete))] // BBFDNCCEJHI
-    public static class VotingComplete
-    {
-        public static void Postfix(MeetingHud __instance)
-        {
-            if (PlayerControl.LocalPlayer.Is(AbilityEnum.Assassin))
-            {
-                var assassin = Ability.GetAbility<Assassin>(PlayerControl.LocalPlayer);
-                ShowHideButtons.HideButtons(assassin);
-            }
-        }
-    }
-}
