+++ source/Patches/NeutralRoles/DoomsayerMod/DoomsayerKill.cs
@@ -1,19 +1,5 @@
 ﻿using System.Linq;
-using Hazel;
 using TownOfUs.Roles;
-using TownOfUs.Roles.Modifiers;
-using UnityEngine;
-using UnityEngine.UI;
-using Il2CppInterop.Runtime.InteropTypes.Arrays;
-using TownOfUs.CrewmateRoles.MedicMod;
-using TownOfUs.Modifiers.AssassinMod;
-using TownOfUs.ImpostorRoles.BlackmailerMod;
-using TownOfUs.Extensions;
-using TownOfUs.CrewmateRoles.VigilanteMod;
-using TownOfUs.CrewmateRoles.SwapperMod;
-using TownOfUs.Patches;
-using Reactor.Utilities.Extensions;
-using TownOfUs.CrewmateRoles.ImitatorMod;
 
 namespace TownOfUs.NeutralRoles.DoomsayerMod
 {
@@ -28,221 +14,17 @@ namespace TownOfUs.NeutralRoles.DoomsayerMod
         }
         public static void RpcMurderPlayer(PlayerVoteArea voteArea, PlayerControl player, PlayerControl doomsayer)
         {
-            DoomKillCount(player, doomsayer);
-            MurderPlayer(voteArea, player);
+            DoomKillCount(doomsayer);
+            (Role.GetRole<Doomsayer>(doomsayer) as IGuesser).MurderPlayer(voteArea, player);
             Utils.Rpc(CustomRPC.DoomsayerKill, player.PlayerId, doomsayer.PlayerId);
         }
 
-        public static void MurderPlayer(PlayerControl player, bool checkLover = true)
-        {
-            PlayerVoteArea voteArea = MeetingHud.Instance.playerStates.First(
-                x => x.TargetPlayerId == player.PlayerId
-            );
-            MurderPlayer(voteArea, player, checkLover);
-        }
-        public static void DoomKillCount(PlayerControl player, PlayerControl doomsayer)
+        public static void DoomKillCount(PlayerControl doomsayer)
         {
             var doom = Role.GetRole<Doomsayer>(doomsayer);
             doom.CorrectAssassinKills += 1;
             doom.GuessedCorrectly += 1;
             if (doom.GuessedCorrectly == CustomGameOptions.DoomsayerGuessesToWin) doom.WonByGuessing = true;
-        }
-        public static void MurderPlayer(
-            PlayerVoteArea voteArea,
-            PlayerControl player,
-            bool checkLover = true
-        )
-        {
-            var hudManager = DestroyableSingleton<HudManager>.Instance;
-            if (checkLover)
-            {
-                SoundManager.Instance.PlaySound(player.KillSfx, false, 0.8f);
-                hudManager.KillOverlay.ShowKillAnimation(player.Data, player.Data);
-            }
-            var amOwner = player.AmOwner;
-            if (amOwner)
-            {
-                Utils.ShowDeadBodies = true;
-                hudManager.ShadowQuad.gameObject.SetActive(false);
-                player.nameText().GetComponent<MeshRenderer>().material.SetInt("_Mask", 0);
-                player.RpcSetScanner(false);
-                ImportantTextTask importantTextTask = new GameObject("_Player").AddComponent<ImportantTextTask>();
-                importantTextTask.transform.SetParent(AmongUsClient.Instance.transform, false);
-                if (!GameOptionsManager.Instance.currentNormalGameOptions.GhostsDoTasks)
-                {
-                    for (int i = 0;i < player.myTasks.Count;i++)
-                    {
-                        PlayerTask playerTask = player.myTasks.ToArray()[i];
-                        playerTask.OnRemove();
-                        Object.Destroy(playerTask.gameObject);
-                    }
-
-                    player.myTasks.Clear();
-                    importantTextTask.Text = DestroyableSingleton<TranslationController>.Instance.GetString(
-                        StringNames.GhostIgnoreTasks,
-                        new Il2CppReferenceArray<Il2CppSystem.Object>(0)
-                    );
-                }
-                else
-                {
-                    importantTextTask.Text = DestroyableSingleton<TranslationController>.Instance.GetString(
-                        StringNames.GhostDoTasks,
-                        new Il2CppReferenceArray<Il2CppSystem.Object>(0));
-                }
-
-                player.myTasks.Insert(0, importantTextTask);
-
-                if (player.Is(RoleEnum.Swapper))
-                {
-                    var swapper = Role.GetRole<Swapper>(PlayerControl.LocalPlayer);
-                    swapper.ListOfActives.Clear();
-                    swapper.Buttons.Clear();
-                    SwapVotes.Swap1 = null;
-                    SwapVotes.Swap2 = null;
-                    Utils.Rpc(CustomRPC.SetSwaps, sbyte.MaxValue, sbyte.MaxValue);
-                    var buttons = Role.GetRole<Swapper>(player).Buttons;
-                    foreach (var button in buttons)
-                    {
-                        button.SetActive(false);
-                        button.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-                    }
-                }
-
-                if (player.Is(RoleEnum.Imitator))
-                {
-                    var imitator = Role.GetRole<Imitator>(PlayerControl.LocalPlayer);
-                    imitator.ListOfActives.Clear();
-                    imitator.Buttons.Clear();
-                    SetImitate.Imitate = null;
-                    var buttons = Role.GetRole<Imitator>(player).Buttons;
-                    foreach (var button in buttons)
-                    {
-                        button.SetActive(false);
-                        button.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-                    }
-                }
-
-                if (player.Is(AbilityEnum.Assassin))
-                {
-                    var assassin = Ability.GetAbility<Assassin>(PlayerControl.LocalPlayer);
-                    ShowHideButtons.HideButtons(assassin);
-                }
-
-                if (player.Is(RoleEnum.Vigilante))
-                {
-                    var vigilante = Role.GetRole<Vigilante>(PlayerControl.LocalPlayer);
-                    ShowHideButtonsVigi.HideButtonsVigi(vigilante);
-                }
-
-                if (player.Is(RoleEnum.Doomsayer))
-                {
-                    var doom = Role.GetRole<Doomsayer>(PlayerControl.LocalPlayer);
-                    ShowHideButtonsDoom.HideButtonsDoom(doom);
-                }
-
-                if (player.Is(RoleEnum.Mayor))
-                {
-                    var mayor = Role.GetRole<Mayor>(PlayerControl.LocalPlayer);
-                    mayor.RevealButton.Destroy();
-                }
-            }
-            player.Die(DeathReason.Kill, false);
-            if (checkLover && player.IsLover() && CustomGameOptions.BothLoversDie)
-            {
-                var otherLover = Modifier.GetModifier<Lover>(player).OtherLover.Player;
-                if (!otherLover.Is(RoleEnum.Pestilence)) MurderPlayer(otherLover, false);
-            }
-
-            var deadPlayer = new DeadPlayer
-            {
-                PlayerId = player.PlayerId,
-                KillerId = player.PlayerId,
-                KillTime = System.DateTime.UtcNow,
-            };
-
-            Murder.KilledPlayers.Add(deadPlayer);
-            if (voteArea == null) return;
-            if (voteArea.DidVote) voteArea.UnsetVote();
-            voteArea.AmDead = true;
-            voteArea.Overlay.gameObject.SetActive(true);
-            voteArea.Overlay.color = Color.white;
-            voteArea.XMark.gameObject.SetActive(true);
-            voteArea.XMark.transform.localScale = Vector3.one;
-
-            var meetingHud = MeetingHud.Instance;
-            if (amOwner)
-            {
-                meetingHud.SetForegroundForDead();
-            }
-
-            var blackmailers = Role.AllRoles.Where(x => x.RoleType == RoleEnum.Blackmailer && x.Player != null).Cast<Blackmailer>();
-            foreach (var role in blackmailers)
-            {
-                if (role.Blackmailed != null && voteArea.TargetPlayerId == role.Blackmailed.PlayerId)
-                {
-                    if (BlackmailMeetingUpdate.PrevXMark != null && BlackmailMeetingUpdate.PrevOverlay != null)
-                    {
-                        voteArea.XMark.sprite = BlackmailMeetingUpdate.PrevXMark;
-                        voteArea.Overlay.sprite = BlackmailMeetingUpdate.PrevOverlay;
-                        voteArea.XMark.transform.localPosition = new Vector3(
-                            voteArea.XMark.transform.localPosition.x - BlackmailMeetingUpdate.LetterXOffset,
-                            voteArea.XMark.transform.localPosition.y - BlackmailMeetingUpdate.LetterYOffset,
-                            voteArea.XMark.transform.localPosition.z);
-                    }
-                }
-            }
-
-            if (PlayerControl.LocalPlayer.Is(RoleEnum.Vigilante) && !PlayerControl.LocalPlayer.Data.IsDead)
-            {
-                var vigi = Role.GetRole<Vigilante>(PlayerControl.LocalPlayer);
-                ShowHideButtonsVigi.HideTarget(vigi, voteArea.TargetPlayerId);
-            }
-
-            if (PlayerControl.LocalPlayer.Is(AbilityEnum.Assassin) && !PlayerControl.LocalPlayer.Data.IsDead)
-            {
-                var assassin = Ability.GetAbility<Assassin>(PlayerControl.LocalPlayer);
-                ShowHideButtons.HideTarget(assassin, voteArea.TargetPlayerId);
-            }
-
-            if (PlayerControl.LocalPlayer.Is(RoleEnum.Doomsayer) && !PlayerControl.LocalPlayer.Data.IsDead)
-            {
-                var doom = Role.GetRole<Doomsayer>(PlayerControl.LocalPlayer);
-                ShowHideButtonsDoom.HideTarget(doom, voteArea.TargetPlayerId);
-            }
-
-            if (PlayerControl.LocalPlayer.Is(RoleEnum.Swapper) && !PlayerControl.LocalPlayer.Data.IsDead)
-            {
-                var swapper = Role.GetRole<Swapper>(PlayerControl.LocalPlayer);
-                var button = swapper.Buttons[voteArea.TargetPlayerId];
-                if (button.GetComponent<SpriteRenderer>().sprite == TownOfUs.SwapperSwitch)
-                {
-                    swapper.ListOfActives[voteArea.TargetPlayerId] = false;
-                    if (SwapVotes.Swap1 == voteArea) SwapVotes.Swap1 = null;
-                    if (SwapVotes.Swap2 == voteArea) SwapVotes.Swap2 = null;
-                    Utils.Rpc(CustomRPC.SetSwaps, sbyte.MaxValue, sbyte.MaxValue);
-                }
-                button.SetActive(false);
-                button.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
-                swapper.Buttons[voteArea.TargetPlayerId] = null;
-            }
-
-            foreach (var playerVoteArea in meetingHud.playerStates)
-            {
-                if (playerVoteArea.VotedFor != player.PlayerId) continue;
-                playerVoteArea.UnsetVote();
-                var voteAreaPlayer = Utils.PlayerById(playerVoteArea.TargetPlayerId);
-                if (voteAreaPlayer.Is(RoleEnum.Prosecutor))
-                {
-                    var pros = Role.GetRole<Prosecutor>(voteAreaPlayer);
-                    pros.ProsecuteThisMeeting = false;
-                }
-                if (!voteAreaPlayer.AmOwner) continue;
-                meetingHud.ClearVote();
-            }
-
-            if (AmongUsClient.Instance.AmHost) meetingHud.CheckForEndVoting();
-
-            AddHauntPatch.AssassinatedPlayers.Add(player);
         }
     }
 }
