--- source/Patches/NeutralRoles/DoomsayerMod/NotVote.cs
@@ -1,18 +1,0 @@
-using HarmonyLib;
-using TownOfUs.Roles;
-
-namespace TownOfUs.NeutralRoles.DoomsayerMod
-{
-    [HarmonyPatch(typeof(MeetingHud), nameof(MeetingHud.VotingComplete))] // BBFDNCCEJHI
-    public static class VotingComplete
-    {
-        public static void Postfix(MeetingHud __instance)
-        {
-            if (PlayerControl.LocalPlayer.Is(RoleEnum.Doomsayer))
-            {
-                var doomsayer = Role.GetRole<Doomsayer>(PlayerControl.LocalPlayer);
-                ShowHideButtonsDoom.HideButtonsDoom(doomsayer);
-            }
-        }
-    }
-}
