+++ source/Patches/Roles/Interfaces/IGuesser.cs
@@ -1,0 +1,339 @@
+using Reactor.Utilities.Extensions;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
+using UnityEngine.UI;
+using TMPro;
+using Object = UnityEngine.Object;
+using TownOfUs.Extensions;
+using TownOfUs.CrewmateRoles.MedicMod;
+using TownOfUs.CrewmateRoles.SwapperMod;
+using TownOfUs.ImpostorRoles.BlackmailerMod;
+using TownOfUs.Roles.Modifiers;
+using Il2CppInterop.Runtime.InteropTypes.Arrays;
+
+namespace TownOfUs.Roles
+{
+    public interface IRoleGuesser<T> : IGuesser where T : Role
+    {
+        T Instance => Role.GetRole<T>(PlayerControl.LocalPlayer);
+    }
+
+    public interface IAbilityGuesser<T> : IGuesser where T : Ability
+    {
+        T Instance => Ability.GetAbility<T>(PlayerControl.LocalPlayer);
+    }
+
+    public interface IGuesser
+    {
+        Dictionary<byte, (GameObject cycleback, GameObject cycleforward, GameObject guess, TMP_Text guess_text)> Buttons { get; set; }
+
+        Dictionary<string, Color> ColorMapping { get; set; }
+
+        Dictionary<string, Color> SortedColorMapping { get; set; }
+
+        Dictionary<byte, string> Guesses { get; set; }
+
+        List<string> PossibleGuesses => SortedColorMapping.Keys.ToList();
+
+        virtual Action Cycle(IGuesser role, PlayerVoteArea voteArea, TextMeshPro nameText, bool forwardsCycle)
+        {
+            void Listener()
+            {
+                if (MeetingHud.Instance.state == MeetingHud.VoteStates.Discussion) return;
+                var currentGuess = role.Guesses[voteArea.TargetPlayerId];
+                var guessIndex = currentGuess == "None"
+                    ? -1
+                    : role.PossibleGuesses.IndexOf(currentGuess);
+                if (forwardsCycle)
+                {
+                    if (++guessIndex >= role.PossibleGuesses.Count)
+                        guessIndex = 0;
+                }
+                else
+                {
+                    if (--guessIndex < 0)
+                        guessIndex = role.PossibleGuesses.Count - 1;
+                }
+
+                var newGuess = role.Guesses[voteArea.TargetPlayerId] = role.PossibleGuesses[guessIndex];
+
+                nameText.text = newGuess == "None"
+                    ? "Guess"
+                    : $"<color=#{role.SortedColorMapping[newGuess].ToHtmlStringRGBA()}>{newGuess}</color>";
+            }
+            return Listener;
+        }
+
+        virtual void GenButton(IGuesser role, PlayerVoteArea voteArea)
+        {
+            var targetId = voteArea.TargetPlayerId;
+            if (role.IsUnGuessable(voteArea))
+            {
+                role.Buttons[targetId] = (null, null, null, null);
+                return;
+            }
+
+            var confirmButton = voteArea.Buttons.transform.GetChild(0).gameObject;
+            var parent = confirmButton.transform.parent.parent;
+
+            var nameText = Object.Instantiate(voteArea.NameText, voteArea.transform);
+            voteArea.NameText.transform.localPosition = new Vector3(0.55f, 0.12f, -0.1f);
+            nameText.transform.localPosition = new Vector3(0.55f, -0.12f, -0.1f);
+            nameText.text = "Guess";
+
+            var cycleBack = Object.Instantiate(confirmButton, voteArea.transform);
+            var cycleRendererBack = cycleBack.GetComponent<SpriteRenderer>();
+            cycleRendererBack.sprite = TownOfUs.CycleBackSprite;
+            cycleBack.transform.localPosition = new Vector3(-0.5f, 0.15f, -2f);
+            cycleBack.transform.localScale = new Vector3(0.8f, 0.8f, 0.8f);
+            cycleBack.layer = 5;
+            cycleBack.transform.parent = parent;
+            var cycleEventBack = new Button.ButtonClickedEvent();
+            cycleEventBack.AddListener(Cycle(role, voteArea, nameText, false));
+            cycleBack.GetComponent<PassiveButton>().OnClick = cycleEventBack;
+            var cycleColliderBack = cycleBack.GetComponent<BoxCollider2D>();
+            cycleColliderBack.size = cycleRendererBack.sprite.bounds.size;
+            cycleColliderBack.offset = Vector2.zero;
+            cycleBack.transform.GetChild(0).gameObject.Destroy();
+
+            var cycleForward = Object.Instantiate(confirmButton, voteArea.transform);
+            var cycleRendererForward = cycleForward.GetComponent<SpriteRenderer>();
+            cycleRendererForward.sprite = TownOfUs.CycleForwardSprite;
+            cycleForward.transform.localPosition = new Vector3(-0.2f, 0.15f, -2f);
+            cycleForward.transform.localScale = new Vector3(0.8f, 0.8f, 0.8f);
+            cycleForward.layer = 5;
+            cycleForward.transform.parent = parent;
+            var cycleEventForward = new Button.ButtonClickedEvent();
+            cycleEventForward.AddListener(Cycle(role, voteArea, nameText, true));
+            cycleForward.GetComponent<PassiveButton>().OnClick = cycleEventForward;
+            var cycleColliderForward = cycleForward.GetComponent<BoxCollider2D>();
+            cycleColliderForward.size = cycleRendererForward.sprite.bounds.size;
+            cycleColliderForward.offset = Vector2.zero;
+            cycleForward.transform.GetChild(0).gameObject.Destroy();
+
+            var guess = Object.Instantiate(confirmButton, voteArea.transform);
+            var guessRenderer = guess.GetComponent<SpriteRenderer>();
+            guessRenderer.sprite = TownOfUs.GuessSprite;
+            guess.transform.localPosition = new Vector3(-0.35f, -0.15f, -2f);
+            guess.transform.localScale = new Vector3(0.8f, 0.8f, 0.8f);
+            guess.layer = 5;
+            guess.transform.parent = parent;
+            var guessEvent = new Button.ButtonClickedEvent();
+            guessEvent.AddListener(Guess(role, voteArea));
+            guess.GetComponent<PassiveButton>().OnClick = guessEvent;
+            var bounds = guess.GetComponent<SpriteRenderer>().bounds;
+            bounds.size = new Vector3(0.52f, 0.3f, 0.16f);
+            var guessCollider = guess.GetComponent<BoxCollider2D>();
+            guessCollider.size = guessRenderer.sprite.bounds.size;
+            guessCollider.offset = Vector2.zero;
+            guess.transform.GetChild(0).gameObject.Destroy();
+
+            role.Guesses.Add(targetId, "None");
+            role.Buttons[targetId] = (cycleBack, cycleForward, guess, nameText);
+        }
+
+        virtual void HideTarget(IGuesser role, byte targetId)
+        {
+            var (cycleBack, cycleForward, guess, guessText) = role.Buttons[targetId];
+            if (cycleBack == null || cycleForward == null) return;
+            cycleBack.SetActive(false);
+            cycleForward.SetActive(false);
+            guess.SetActive(false);
+            guessText.gameObject.SetActive(false);
+
+            cycleBack.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+            cycleForward.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+            guess.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+            role.Buttons[targetId] = (null, null, null, null);
+            role.Guesses.Remove(targetId);
+        }
+
+        virtual void HideAllButtons(IGuesser role)
+        {
+            foreach (var (_, (cycleBack, cycleForward, guess, guessText)) in role.Buttons)
+            {
+                if (cycleBack == null || cycleForward == null) continue;
+                cycleBack.SetActive(false);
+                cycleForward.SetActive(false);
+                guess.SetActive(false);
+                guessText.gameObject.SetActive(false);
+
+                cycleBack.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+                cycleForward.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+                guess.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+            }
+        }
+
+        virtual void MurderPlayer(PlayerControl player, bool checkLover = true)
+        {
+            PlayerVoteArea voteArea = MeetingHud.Instance.playerStates.First(
+                x => x.TargetPlayerId == player.PlayerId
+            );
+            MurderPlayer(voteArea, player, checkLover);
+        }
+
+        virtual void MurderPlayer(PlayerVoteArea voteArea, PlayerControl player, bool checkLover = true)
+        {
+            var hudManager = DestroyableSingleton<HudManager>.Instance;
+            if (checkLover)
+            {
+                SoundManager.Instance.PlaySound(player.KillSfx, false, 0.8f);
+                hudManager.KillOverlay.ShowKillAnimation(player.Data, player.Data);
+            }
+            var amOwner = player.AmOwner;
+            if (amOwner)
+            {
+                Utils.ShowDeadBodies = true;
+                hudManager.ShadowQuad.gameObject.SetActive(false);
+                player.nameText().GetComponent<MeshRenderer>().material.SetInt("_Mask", 0);
+                player.RpcSetScanner(false);
+                ImportantTextTask importantTextTask = new GameObject("_Player").AddComponent<ImportantTextTask>();
+                importantTextTask.transform.SetParent(AmongUsClient.Instance.transform, false);
+                if (!GameOptionsManager.Instance.currentNormalGameOptions.GhostsDoTasks)
+                {
+                    for (int i = 0;i < player.myTasks.Count;i++)
+                    {
+                        PlayerTask playerTask = player.myTasks.ToArray()[i];
+                        playerTask.OnRemove();
+                        Object.Destroy(playerTask.gameObject);
+                    }
+
+                    player.myTasks.Clear();
+                    importantTextTask.Text = DestroyableSingleton<TranslationController>.Instance.GetString(
+                        StringNames.GhostIgnoreTasks,
+                        new Il2CppReferenceArray<Il2CppSystem.Object>(0)
+                    );
+                }
+                else
+                {
+                    importantTextTask.Text = DestroyableSingleton<TranslationController>.Instance.GetString(
+                        StringNames.GhostDoTasks,
+                        new Il2CppReferenceArray<Il2CppSystem.Object>(0));
+                }
+
+                player.myTasks.Insert(0, importantTextTask);
+
+                if (player.Is(AbilityEnum.Assassin))
+                {
+                    var assassin = Ability.GetAbility<Assassin>(PlayerControl.LocalPlayer) as IGuesser;
+                    assassin.HideAllButtons(assassin);
+                }
+
+                if (player.Is(RoleEnum.Doomsayer))
+                {
+                    var doomsayer = Role.GetRole<Doomsayer>(PlayerControl.LocalPlayer) as IGuesser;
+                    doomsayer.HideAllButtons(doomsayer);
+                }
+
+                if (player.Is(RoleEnum.Vigilante))
+                {
+                    var retributionist = Role.GetRole<Vigilante>(PlayerControl.LocalPlayer) as IGuesser;
+                    retributionist.HideAllButtons(retributionist);
+                }
+            }
+            player.Die(DeathReason.Kill, false);
+            if (checkLover && player.IsLover() && CustomGameOptions.BothLoversDie)
+            {
+                var otherLover = Modifier.GetModifier<Lover>(player).OtherLover.Player;
+                if (!otherLover.Is(RoleEnum.Pestilence)) MurderPlayer(otherLover, false);
+            }
+
+            var deadPlayer = new DeadPlayer
+            {
+                PlayerId = player.PlayerId,
+                KillerId = player.PlayerId,
+                KillTime = DateTime.UtcNow,
+            };
+
+            Murder.KilledPlayers.Add(deadPlayer);
+            if (voteArea == null) return;
+            if (voteArea.DidVote) voteArea.UnsetVote();
+            voteArea.AmDead = true;
+            voteArea.Overlay.gameObject.SetActive(true);
+            voteArea.Overlay.color = Color.white;
+            voteArea.XMark.gameObject.SetActive(true);
+            voteArea.XMark.transform.localScale = Vector3.one;
+
+            var meetingHud = MeetingHud.Instance;
+            if (amOwner)
+            {
+                meetingHud.SetForegroundForDead();
+            }
+
+            var blackmailers = Role.AllRoles.Where(x => x.RoleType == RoleEnum.Blackmailer && x.Player != null).Cast<Blackmailer>();
+            foreach (var role in blackmailers)
+            {
+                if (role.Blackmailed != null && voteArea.TargetPlayerId == role.Blackmailed.PlayerId)
+                {
+                    if (BlackmailMeetingUpdate.PrevXMark != null && BlackmailMeetingUpdate.PrevOverlay != null)
+                    {
+                        voteArea.XMark.sprite = BlackmailMeetingUpdate.PrevXMark;
+                        voteArea.Overlay.sprite = BlackmailMeetingUpdate.PrevOverlay;
+                        voteArea.XMark.transform.localPosition = new Vector3(
+                            voteArea.XMark.transform.localPosition.x - BlackmailMeetingUpdate.LetterXOffset,
+                            voteArea.XMark.transform.localPosition.y - BlackmailMeetingUpdate.LetterYOffset,
+                            voteArea.XMark.transform.localPosition.z);
+                    }
+                }
+            }
+
+            if (PlayerControl.LocalPlayer.Is(RoleEnum.Vigilante) && !PlayerControl.LocalPlayer.Data.IsDead)
+            {
+                var vigi = Role.GetRole<Vigilante>(PlayerControl.LocalPlayer) as IGuesser;
+                vigi.HideTarget(vigi, voteArea.TargetPlayerId);
+            }
+
+            if (PlayerControl.LocalPlayer.Is(AbilityEnum.Assassin) && !PlayerControl.LocalPlayer.Data.IsDead)
+            {
+                var assassin = Ability.GetAbility<Assassin>(PlayerControl.LocalPlayer) as IGuesser;
+                assassin.HideTarget(assassin, voteArea.TargetPlayerId);
+            }
+
+            if (PlayerControl.LocalPlayer.Is(RoleEnum.Doomsayer) && !PlayerControl.LocalPlayer.Data.IsDead)
+            {
+                var doom = Role.GetRole<Doomsayer>(PlayerControl.LocalPlayer) as IGuesser;
+                doom.HideTarget(doom, voteArea.TargetPlayerId);
+            }
+
+            if (PlayerControl.LocalPlayer.Is(RoleEnum.Swapper) && !PlayerControl.LocalPlayer.Data.IsDead)
+            {
+                var swapper = Role.GetRole<Swapper>(PlayerControl.LocalPlayer);
+                var button = swapper.Buttons[voteArea.TargetPlayerId];
+                if (button.GetComponent<SpriteRenderer>().sprite == TownOfUs.SwapperSwitch)
+                {
+                    swapper.ListOfActives[voteArea.TargetPlayerId] = false;
+                    if (SwapVotes.Swap1 == voteArea) SwapVotes.Swap1 = null;
+                    if (SwapVotes.Swap2 == voteArea) SwapVotes.Swap2 = null;
+                    Utils.Rpc(CustomRPC.SetSwaps, sbyte.MaxValue, sbyte.MaxValue);
+                }
+                button.SetActive(false);
+                button.GetComponent<PassiveButton>().OnClick = new Button.ButtonClickedEvent();
+                swapper.Buttons[voteArea.TargetPlayerId] = null;
+            }
+
+            foreach (var playerVoteArea in meetingHud.playerStates)
+            {
+                if (playerVoteArea.VotedFor != player.PlayerId) continue;
+                playerVoteArea.UnsetVote();
+                var voteAreaPlayer = Utils.PlayerById(playerVoteArea.TargetPlayerId);
+                if (voteAreaPlayer.Is(RoleEnum.Prosecutor))
+                {
+                    var pros = Role.GetRole<Prosecutor>(voteAreaPlayer);
+                    pros.ProsecuteThisMeeting = false;
+                }
+                if (!voteAreaPlayer.AmOwner) continue;
+                meetingHud.ClearVote();
+            }
+
+            if (AmongUsClient.Instance.AmHost) meetingHud.CheckForEndVoting();
+
+            Patches.AddHauntPatch.AssassinatedPlayers.Add(player);
+        }     
+        
+        Action Guess(IGuesser role, PlayerVoteArea voteArea);
+
+        bool IsUnGuessable(PlayerVoteArea voteArea);
+    }
+}
